// src/types/workout.types.ts
export type Difficulty = "Beginner" | "Intermediate" | "Advanced";


export interface Workout {
    id: number;
    name: string;
    image: string;
    muscleGroups: string[];
    equipment: string;
    difficulty: Difficulty;
    duration: number;
    caloriesBurned: number;
    sets: number;
    reps: string;
    rating: number;
    description: string;
    instructions: string[];
}